Chat_bot-main/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
huggingface-hub
sentence-transformers